pyproject.toml
[tool.poetry]
name = "fastapi-social-network"
version = "0.1.0"
description = "A project for getting to know FastAPI."
authors = ["None <name@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
SQLAlchemy = "^1.4.26"
psycopg2 = "^2.9.1"
email-validator = "^1.1.3"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
python-jose = "^3.3.0"
cryptography = "^35.0.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
flake8 = "^4.0.1"
flake8-bugbear = "^21.9.2"
flake8-builtins = "^1.5.3"
flake8-comprehensions = "^3.7.0"
pep8-naming = "^0.12.1"

[virtualenvs]
create = true
in-project = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120